Fixed splash screen configuration (#340)
# Changes
Simplified the splash screen setup, which also fixes an issue on
Android, where the icon would be very small on HDPI screens. It is now
set up in the way recommended by the official expo guide.

| **Before** | **After** |
| ----------------- | ------------- |
|
[before.webm](https://github.com/user-attachments/assets/a0f480ad-64d8-4f14-bce9-46763fb9d688)
|
[after.webm](https://github.com/user-attachments/assets/30cf2e2a-6a17-4598-a93a-ad50d5f02e7e)
|

# Blockers
Please test iOS before merging, I haven't tested iOS. This has to be
tested in release mode, since the splash screen is managed by expo in
debug mode (on Android I had to manually build the release APK and then
install it on the test device).

# Related issues
Closes #318